+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 -   Support for reddit galleries ([#312](https://github.com/Tunous/Dawn/pull/312))
 
+### Fixed
+
+-   Correctly parse links with leading/trailing spaces ([319](https://github.com/Tunous/Dawn/pull/319))
+
 ## [0.10.0] - 2020-08-14
 
 ### Added